+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Change `pageview_id`/`visit_id` format
 * Add support for Edge integration and deprecated `Composer.accessToken`
 * Fix missing browserID at loading templates
+* Clean up deprecated APIs 
 * Updated dependencies:
     - androidx.fragment:fragment [1.6.1 -> 1.6.2]
       https://developer.android.com/jetpack/androidx/releases/fragment#1.6.2